+++ plugins/woocommerce-admin/bin/update-version.php
@@ -11,7 +11,7 @@ $package      = json_decode( $package_json );
 function replace_version( $filename, $package_json ) {
 	$lines = array();
 	$file  = file( $filename );
-	
+
 	foreach ( $file as $line ) {
 		if ( stripos( $line, ' * Version: ' ) !== false ) {
 			$line = " * Version: {$package_json->version}\n";